+++ src/api/admin.py
@@ -1,7 +1,7 @@
 import os
 from flask_admin import Admin
 from flask_admin.contrib.sqla import ModelView
-from .models import db, Users
+from .models import db, Users, Activities , ActivitiesHistory
 
 def setup_admin(app):
     app.secret_key = os.environ.get('FLASK_APP_KEY', 'sample key')
@@ -9,3 +9,5 @@ def setup_admin(app):
     admin = Admin(app, name='4Geeks Admin', template_mode='bootstrap3')
     # Add your models here, for example this is how we add a the User model to the admin
     admin.add_view(ModelView(Users, db.session))  # You can duplicate that line to add mew models
+    admin.add_view(ModelView(Activities, db.session))
+    admin.add_view(ModelView(ActivitiesHistory, db.session))
